+++ node/src/types/kernel/api/system/getConf/response.d.ts
@@ -70,7 +70,7 @@ export interface IConf {
      */
     lang:     TLang;
     langs:    any;
-    localIPS: any;
+    localIPs: any;
     /**
      * Log level
      */